requirements.txt
argparse==1.4.0
blessings==1.6
bpython==0.15
celery==3.1.18
cffi==1.5.2
configparser==3.5.0
cryptography==1.2.2
cssselect==0.9.1
curtsies==0.2.6
django==1.4.2
django-composite-foreignkey==1.0.0a10
django-debug-toolbar==0.9.4
django-mptt==0.6.1
django-redis==3.8.4
enum==0.4.6
greenlet==0.4.9
html5lib==0.9999999
importlib==1.0.4
ipaddr==2.1.11
itsdangerous==0.24
linecache2==1.0.0
lxml==3.5.0
m2crypto==0.25.1
manifestparser==1.1
marionette-client==3.1.0
misaka==1.0.2
mozcrash==0.17
mozdevice==0.48
mozinfo==0.9
mozlog==3.3
moznetwork==0.27
mozprocess==0.23
mozprofile==0.28
mozrunner==6.12
moztest==0.7
mozversion==1.4
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
ordereddict==1.1
Pillow==6.0.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyDes==2.0.1
pyexcelerator==0.6.4.1
pygments==2.1.1
pyopenssl==0.15.1
pyparsing==2.1.4
pyquery==1.2.11
pyreadline==2.1
python-dateutil==2.1
qrcode==5.1
redis==2.10.5
requests==2.8.1
simplejson==3.8.1
six==1.10.0
sphinx-me==0.3
traceback2==1.4.0
unittest2==1.1.0
wcwidth==0.1.6
web.py==0.40.dev0
werkzeug==0.11.11
wptserve==1.4.0
xlwt==1.3.0
xmltodict==0.9.2
